% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Academic_year.R
\name{academic_year}
\alias{academic_year}
\title{Academic year}
\usage{
academic_year(x, start_1_oct = FALSE)
}
\arguments{
\item{x}{A date, or vector with multiple dates. POSIXct is also accepted.}

\item{start_1_oct}{Does the academic year start on the 1st of October? default FALSE:
based on September 1st}
}
\value{
The academic year in which the specified date falls
}
\description{
In this function, a date is translated to the academic year in which it
falls. This is based on a start of the academic year on the
1st of September.
}
\examples{
academic_year(lubridate::today())

}
\seealso{
Other vector calculations: 
\code{\link{clean_multiple_underscores}()},
\code{\link{interval_round}()},
\code{\link{sum_0_1}()},
\code{\link{transform_01_to_ft}()}
}
\concept{vector calculations}
