% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_multiple_underscores.R
\name{clean_multiple_underscores}
\alias{clean_multiple_underscores}
\title{clean multiple underscores}
\usage{
clean_multiple_underscores(x)
}
\arguments{
\item{x}{The vector or string to be cleaned.}
}
\value{
cleaned vector or string.
}
\description{
Replaces multiple underscores into a single underscore in a vector or string.
}
\examples{
clean_multiple_underscores("hello___world")

}
\seealso{
Other vector calculations: 
\code{\link{academic_year}()},
\code{\link{interval_round}()},
\code{\link{sum_0_1}()},
\code{\link{transform_01_to_ft}()}
}
\concept{vector calculations}
