% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary.R
\name{group_summary}
\alias{group_summary}
\title{Group Summary}
\usage{
group_summary(data, group_var, group_func = mean)
}
\arguments{
\item{data}{Input data source.}

\item{group_var}{Variable to make the group by.}

\item{group_func}{Function to be used in the group by. Default is mean.}
}
\value{
Grouped data frame.
}
\description{
Calculate the means (or other function) per group to analyze how each segment behaves. It scales each variable mean into the 0 to 1 range to easily profile the groups according to its mean. It also calculates the mean regardless of the grouping. This function is also useful when you want to profile cluster results in terms of its means. It automatically adds a row representing the summary of the column regardless of the group_var categories, which is useful to compare each segment with the whole population. It will exclude all factor/character variables.
}
