% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary_rank.R
\name{group_summary_rank}
\alias{group_summary_rank}
\title{Group Summary Rank}
\usage{
group_summary_rank(data, group_var, group_func = mean)
}
\arguments{
\item{data}{Input data source.}

\item{group_var}{Variable to make the group by.}

\item{group_func}{Function to be used in the group by. Default is mean.}
}
\value{
Grouped data frame, showing the rank instead of the absolute values.
}
\description{
Similar to 'group_summary' function, this one computes the rank of each value in order to quickly know what is the value in each segment that has the highest value (rank=1). 1 represents the highest number. It will exclude all factor/character variables.
}
