% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interval_round.R
\name{interval_round}
\alias{interval_round}
\title{Interval round}
\usage{
interval_round(x, interval)
}
\arguments{
\item{x}{The numeric vector to adjust}

\item{interval}{The interval sequence}
}
\value{
The vector corrected for the given interval
}
\description{
Function to round numeric values in a vector to values from
an interval sequence.
}
\examples{
interval_round(c(5,4,2,6), interval = seq(1:4))

}
\seealso{
Other vector calculations: 
\code{\link{academic_year}()},
\code{\link{clean_multiple_underscores}()},
\code{\link{sum_0_1}()},
\code{\link{transform_01_to_ft}()}
}
\concept{vector calculations}
