% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{mode}
\alias{mode}
\title{Mode (most common value)}
\usage{
mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{If TRUE: Remove nas before the calculation is done}
}
\value{
the most common value in the vector x
}
\description{
Determine the most common value in a vector. If two values have the same frequency,
the first occurring value is used.
}
\examples{
mode(c(0,3,5,7,5,3,2))


}
