% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_name.R
\name{month_name}
\alias{month_name}
\title{Month Name}
\usage{
month_name(month_numeric, lang = "nl")
}
\arguments{
\item{month_numeric}{Numeric in range 1 - 12.}

\item{lang}{The language of the month names. Default is "nl" (Dutch).}
}
\value{
Character string representation of month in specified language.
}
\description{
Transform month from numeric to equivalent in specified language.
}
\seealso{
Other vector calculations: 
\code{\link{academic_year}()},
\code{\link{clean_multiple_underscores}()},
\code{\link{interval_round}()},
\code{\link{sum_0_1}()},
\code{\link{transform_01_to_ft}()}
}
\concept{vector calculations}
