% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum_0_1.R
\name{sum_0_1}
\alias{sum_0_1}
\title{Sum 0 1}
\usage{
sum_0_1(x)
}
\arguments{
\item{x}{a vector with numeric values}
}
\value{
0 or 1. Depending on whether the sum is greater than 0 or not.
}
\description{
This function is the same as sum(), with one exception: If the outcome value is higher
than 1, it will always return 1.
}
\seealso{
Other vector calculations: 
\code{\link{academic_year}()},
\code{\link{clean_multiple_underscores}()},
\code{\link{interval_round}()},
\code{\link{transform_01_to_ft}()}
}
\concept{vector calculations}
