% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_01_to_ft_new.R
\name{t_f_maker}
\alias{t_f_maker}
\title{T_F_maker}
\usage{
t_f_maker(x)
}
\arguments{
\item{x}{Vector containg zeros and ones.}
}
\description{
Deprecated function to transform vector to boolean.
}
\seealso{
Other vector calculations: 
\code{\link{academic_year}()},
\code{\link{clean_multiple_underscores}()},
\code{\link{interval_round}()},
\code{\link{sum_0_1}()},
\code{\link{transform_01_to_ft}()}

Other booleans: 
\code{\link{test_01}()},
\code{\link{transform_01_to_ft}()}
}
\concept{booleans}
\concept{vector calculations}
