% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_01_to_FT.R
\name{test_01}
\alias{test_01}
\title{Test 01}
\usage{
test_01(x)
}
\arguments{
\item{x}{The vector to test}
}
\value{
A TRUE/FALSE value on the test
}
\description{
This function tests whether the vector is actually a boolean, but is encoded
as a 0/1 variable. The function checks for numeric vectors
whether the only occurring values are 0, 1, or NA. At character and factor
vectors checks whether the only occurring values are "0", "1", or NA
to be.
If there is a 0/1 variable, TRUE is returned, in all others
cases FALSE.
}
\examples{
vector <- c(0, 1, 0, 1, 1, 1, 0)
test_01(vector)
}
\seealso{
Other booleans: 
\code{\link{transform_01_to_ft}()}
}
\concept{booleans}
\concept{tests}
