% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_yes_no.R
\name{test_yes_no}
\alias{test_yes_no}
\title{Test Yes/No Responses}
\usage{
test_yes_no(responses)
}
\arguments{
\item{responses}{A vector of responses.}
}
\value{
A logical vector indicating if each response is yes or no.
}
\description{
This function tests if a vector of responses are yes or no.
}
