% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_logical_yes_no.R
\name{transform_logical_yes_no}
\alias{transform_logical_yes_no}
\title{Transform Logical to Yes/No and Vice Versa}
\usage{
transform_logical_yes_no(x, lang = "nl")
}
\arguments{
\item{x}{A logical or character vector.}

\item{lang}{The language of the yes/no strings. Default is "nl" (Dutch).}
}
\value{
A vector of yes/no strings or a logical vector.
}
\description{
This function transforms a logical vector to a vector of yes/no strings or vice versa.
}
