% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_yes_no.R
\name{translate_yes_no}
\alias{translate_yes_no}
\title{Translate Yes/No Responses}
\usage{
translate_yes_no(responses, source_language = "nl")
}
\arguments{
\item{responses}{A vector of responses.}

\item{source_language}{The language of the responses. Default is "nl" (Dutch).}
}
\value{
A vector of translated responses.
}
\description{
This function translates yes/no responses from a given language to English.
}
