% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim}
\usage{
trim(x)
}
\arguments{
\item{x}{A text string.}
}
\value{
Cleaned string.
}
\description{
Trim both leading and trailing whitespaces from string.
}
\examples{
trim(" hello ")

}
