% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_top_10.R
\name{median_top_10}
\alias{median_top_10}
\title{Median top 10 percentage}
\usage{
median_top_10(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numerical vector}

\item{na.rm}{Default TRUE: Remove NAs, before calculations.}
}
\value{
A numerical value
}
\description{
Calculate the median of the top ten percentage of the values.
}
\examples{
median_top_10(mtcars$cyl)


}
