% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_all_in_file.R
\name{str_replace_all_in_file}
\alias{str_replace_all_in_file}
\title{Replace all occurences of a pattern in a file}
\usage{
str_replace_all_in_file(
  file,
  pattern,
  replacement = "[...]",
  only_comments = TRUE,
  collapse = FALSE
)
}
\arguments{
\item{file}{character, path of file to be modified}

\item{pattern}{character, pattern to be replaced}

\item{replacement}{character, replacement text}

\item{only_comments}{logical, should the replacement only be done in comments}

\item{collapse}{logical, should the lines be collapsed into a single line before replacement}
}
\value{
NULL, the file is modified in place
}
\description{
Replace all occurences of a pattern in a file
}
